Fix mart sales bug and Refactor file paths and components for sales page
