+++ src/components/pages/marts/products/sales/formik-wrapper/components/create-sale-form-wrapper/components/create-sale-form/components/cashable-uuid-field-component.tsx
