+++ src/components/pages/marts/products/sales/formik-wrapper/components/create-sale-form-wrapper/components/create-sale-form/components/details-field-component/index.tsx
