+++ src/components/pages/marts/products/sales/formik-wrapper/components/create-sale-form-wrapper/index.tsx
@@ -1,0 +1,6 @@
+import { memo } from 'react'
+import { CreateSaleFormWrapper as main } from './create-sale-form-wrapper'
+
+const CreateSaleFormWrapper = memo(main)
+
+export default CreateSaleFormWrapper
